+++ packages/error-codes/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @module-federation/error-codes
 
+## 2.3.3
+
 ## 2.3.2
 
 ## 2.3.1