+++ packages/manifest/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @module-federation/manifest
 
+## 2.3.3
+
+### Patch Changes
+
+- Updated dependencies [6d33275]
+  - @module-federation/dts-plugin@2.3.3
+  - @module-federation/sdk@2.3.3
+  - @module-federation/managers@2.3.3
+
 ## 2.3.2
 
 ### Patch Changes